requirements.txt
git+https://github.com/Rapptz/discord.py
requests
aiohttp
pyfiglet
owotext
dadjokes
git+https://github.com/Gorialis/jishaku@master
pymongo
dnspython
psutil
git+https://github.com/LiBa001/disputils
git+https://github.com/ytdl-org/youtube-dl
google-api-python-client
PyNaCl
chatbotAI
topggpy
validators
motor
dblpy
twitchAPI
humanfriendly
pillow
typing
DiscordUtils

git+https://github.com/nirlep5252/Discord-Games.git
git+https://github.com/nirlep5252/discord-ext-events

git+https://github.com/Nirlep5252/epicbot-images/